pyproject.toml
[project]
name = "careerpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "en-core-web-sm",
    "fastapi>=0.136.0",
    "kaggle>=2.0.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pandoc>=2.4",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.2",
    "pypandoc>=1.17",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
    "spacy>=3.8.14",
    "tiktoken>=0.12.0",
    "torch==2.10.0+cu130",
    "torchvision==0.25.0+cu130",
    "transformers>=5.5.4",
    "uvicorn>=0.44.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
